backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
aiofiles==23.2.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3